:recycle: Moved some files to more logical places
